Revert "refactor: modified save & create items logic"
