+++ src/main/java/kr/or/cola/backend/todo/TodoService.java
@@ -7,7 +7,7 @@ import kr.or.cola.backend.todo.dto.TodoResponseDto;
 import kr.or.cola.backend.todo.folder.domain.Folder;
 import kr.or.cola.backend.todo.folder.domain.FolderRepository;
 import kr.or.cola.backend.todo.item.domain.Progress;
-import kr.or.cola.backend.todo.item.dto.ItemsResponseDto;
+import kr.or.cola.backend.todo.folder.dto.ItemsResponseDto;
 import kr.or.cola.backend.todo.item.domain.Item;
 import kr.or.cola.backend.todo.item.domain.ItemRepository;
 import kr.or.cola.backend.user.domain.User;
