--- src/main/java/kr/or/cola/backend/todo/dto/TodoSaveRequestDto.java
@@ -1,4 +1,0 @@
-package kr.or.cola.backend.todo.dto;
-
-public class TodoSaveRequestDto {
-}
