+++ src/main/java/kr/or/cola/backend/todo/item/domain/ItemRepository.java
@@ -5,12 +5,11 @@ import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface ItemRepository extends JpaRepository<Item, Long> {
     List<Item> findByDate(LocalDate date);
-    Optional<Item> findByDateAndFolderId(LocalDate date, Long folderId);
+    Item findByDateAndFolderId(LocalDate date, Long folderId);
     List<Item> findAllByDateAndFolderIdIn(LocalDate date, List<Long> folders);
     void deleteAllByFolderId(Long folderId);
